pyproject.toml
[project]
name = "trimmly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Guilherme Rodrigues", email = "guilherme.b.rodrigues@skf.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.116.1",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
